--- src/main/java/model/Student.java
@@ -1,30 +1,0 @@
-package model;
-
-public class Student {
-    int id;
-    String name;
-
-    public Student() {
-    }
-
-    public Student(int id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
