+++ src/main/resources/application.properties
@@ -1,1 +1,18 @@
+spring.datasource.url=jdbc:mysql://localhost:3306/test-student-B?createDatabaseIfNotExist=true
+spring.datasource.username=root
+spring.datasource.password=15061993
+#spring.jpa.show-sql=true
+#server.port=8080
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialect
+
+
+
+
+
+
+
+
+
